Move testcase in its own file
